Crate transaction_signing

Structs

SignatureAndChecksum

Enums

Error

Transaction signing error.

SignatureType

SufficientContent

Possible content to generate sufficient crypto for

Functions

create_signature

Function to create signatures using RN output action line, and user entered pin and password. Also needs database name to fetch saved transaction and key.

handle_sign

handle_stub

sign_content

Type Aliases

Result

Transaction signing result.