Crate transaction_signing

Re-exports

pub use penumbra::derive_spend_auth_key as penumbra_derive_spend_auth_key;

pub use penumbra::sign_spend as penumbra_sign_spend;

pub use penumbra::sign_transaction as penumbra_sign_transaction;

pub use penumbra::EffectHash as PenumbraEffectHash;

pub use penumbra::PenumbraAuthorizationData;

pub use penumbra::SpendKeyBytes as PenumbraSpendKeyBytes;

pub use penumbra::PENUMBRA_BIP44_PATH;

pub use penumbra::PENUMBRA_COIN_TYPE;

pub use cosmos::sign_cosmos_amino;

pub use cosmos::CosmosSignDocDisplay;

pub use cosmos::CosmosSignRequest;

pub use zcash::decode_notes_bundle_from_cbor;

pub use zcash::derive_orchard_fvk as zcash_derive_orchard_fvk;

pub use zcash::derive_transparent_address as zcash_derive_transparent_address;

pub use zcash::encode_notes_bundle_to_cbor;

pub use zcash::sign_orchard_action as zcash_sign_orchard;

pub use zcash::sign_pczt as zcash_sign_pczt;

pub use zcash::sign_transparent as zcash_sign_transparent;

pub use zcash::verify_anchor_attestation;

pub use zcash::verify_merkle_path;

pub use zcash::OrchardFullViewingKey as ZcashOrchardFvk;

pub use zcash::OrchardSpendingKey as ZcashOrchardKey;

pub use zcash::PcztSignerInput;

pub use zcash::PcztSignerOutput;

pub use zcash::TransparentSpendingKey as ZcashTransparentKey;

pub use zcash::ZcashAuthorizationData;

pub use zcash::ZcashFvkExportData;

pub use zcash::ZcashNoteSyncResult;

pub use zcash::ZcashNoteWithPath;

pub use zcash::ZcashNotesBundle;

pub use zcash::ZcashSignRequest;

pub use zcash::ZcashSignatureResponse;

pub use zcash::ZcashVerifiedNote;

pub use zcash::QR_TYPE_ZCASH_FVK_EXPORT;

pub use zcash::QR_TYPE_ZCASH_NOTES;

pub use zcash::QR_TYPE_ZCASH_SIGNATURES;

pub use zcash::QR_TYPE_ZCASH_SIGN_REQUEST;

pub use zcash::ZCASH_COIN_TYPE;

Modules

cosmos

cosmos transaction signing

penumbra

penumbra transaction signing

zcash

zcash transaction signing

Structs

SignatureAndChecksum

Enums

Error

Transaction signing error.

SignatureType

SufficientContent

Possible content to generate sufficient crypto for

Functions

attestation_digest

Re-export the attestation digest builder from frost-spend. Compute the attestation digest: SHA-256 of domain-separated attestation data.

create_signature

Function to create signatures using RN output action line, and user entered pin and password. Also needs database name to fetch saved transaction and key.

handle_sign

handle_stub

sign_content

Type Aliases

Result

Transaction signing result.